WEBAPP-145: Refactor events route
