+++ src/routes/events/components/EventDetail.css
