+++ src/routes/events/components/EventList.css
@@ -1,46 +1,7 @@
 @import url('../../../modules/common/constants/helpers.css');
-@import url('../../../modules/common/constants/colors.css');
 
 .list a {
   @mixin remove-a;
-}
-
-.first-event {
-  border-top: 2px solid var(--theme-color);
-}
-
-.event {
-  display: flex;
-  border-bottom: 2px solid var(--theme-color);
-}
-
-.event-thumbnail {
-  width: 100px;
-  height: 100px;
-  flex-shrink: 0;
-  padding: 15px 5px;
-}
-
-.event-description {
-  height: 100%;
-  min-width: 1px; /* needed to enable line breaks for too long words, exact value doesn't matter */
-  flex-grow: 1;
-  margin-left: 15px;
-  padding: 15px 0;
-  word-wrap: break-word;
-}
-
-.event-title {
-  font-weight: 900;
-}
-
-.event-date {
-  padding-bottom: 10px;
-  font-weight: 200;
-}
-
-.horizontal-line {
-  border-bottom: 2px solid var(--theme-color);
 }
 
 .no-events {
